fix: coderabbit feedback
