+++ .env.example
@@ -2,6 +2,7 @@
 APP_PORT=8000
 APP_ENV=development
 LOG_LEVEL=debug
+PUBLIC_APP_URL=
 
 # === Database (PostgreSQL) ===
 DATABASE_URL=postgresql://brewform:brewform@localhost:5432/brewform?connection_limit=10&pool_timeout=30
