+++ .eser/specs/verify-fix-inline-findings-code-review/progress.json
@@ -1,0 +1,35 @@
+{
+  "spec": "verify-fix-inline-findings-code-review",
+  "status": "executing",
+  "tasks": [
+    {
+      "id": "task-1",
+      "title": "Every finding is verified against current code, fixed only when still valid, skipped with a brief documented reason, and the entire chang...",
+      "status": "pending"
+    },
+    {
+      "id": "task-2",
+      "title": "Validation plan: 1) Run `deno check --unstable-sloppy-imports apps/api/src/main.ts` to verify type safety after imports move. 2) Run `deno lint apps/ packages/` to catch style issues. 3) Run `deno test apps/api/src/ packages/shared/src/` to confirm tests pass (including the updated share.test.ts import). 4) Run `deno install` after package.json change and verify deno.lock updates cleanly. 5) Manual review of deployment_plan.md markdown with a linter (MD040, MD028).",
+      "status": "pending"
+    },
+    {
+      "id": "task-3",
+      "title": "Write or update tests for all new and changed behavior",
+      "status": "pending"
+    },
+    {
+      "id": "task-4",
+      "title": "Update documentation for all public-facing changes (README, API docs, CHANGELOG)",
+      "status": "pending"
+    }
+  ],
+  "decisions": [
+    {
+      "question": "Split spec into separate areas?",
+      "choice": "Chose to keep as single spec despite multiple areas detected",
+      "promoted": false
+    }
+  ],
+  "debt": [],
+  "updatedAt": "2026-05-06T10:13:16.019Z"
+}
