+++ apps/api/package.json
@@ -20,10 +20,10 @@
     "date-fns": "^4.1.0",
     "bcryptjs": "^2.4.3",
     "qrcode": "^1.5.4",
-    "nodemailer": "^7.0.0",
-    "mjml": "^4.15.0"
+    "nodemailer": "^7.0.0"
   },
   "devDependencies": {
+    "mjml": "^4.15.0",
     "@types/bcryptjs": "^2.4.6",
     "@types/qrcode": "^1.5.5",
     "@types/nodemailer": "^6.4.17"
