+++ apps/api/src/routes/share.test.ts
@@ -1,6 +1,6 @@
 import { describe, it } from 'jsr:@std/testing/bdd';
 import { expect } from 'jsr:@std/expect';
-import { escapeHtml } from './share.ts';
+import { escapeHtml } from '@brewform/shared/utils';
 
 describe('Share Route', () => {
   describe('escapeHtml', () => {
