+++ apps/api/src/routes/share.ts
@@ -1,8 +1,12 @@
 import { Hono } from 'hono';
+import { escapeHtml, escapeHtmlAttr } from '@brewform/shared/utils';
 import { getRecipeMeta } from '../modules/recipe/service.ts';
 import type { AppEnv } from '../types/hono.ts';
 
 const share = new Hono<AppEnv>();
+
+const RECIPE_NOT_FOUND_HTML =
+  '<!DOCTYPE html><html><head><title>Not Found</title></head><body><h1>404 — Recipe not found</h1></body></html>';
 
 const OG_TEMPLATE = (meta: {
   title: string;
@@ -20,41 +24,28 @@ const OG_TEMPLATE = (meta: {
   <meta property="og:title" content="${escapeHtml(meta.title)}">
   <meta property="og:description" content="${escapeHtml(meta.description)}">
   <meta property="og:type" content="article">
-  <meta property="og:url" content="${escapeHtml(meta.url)}">
+  <meta property="og:url" content="${escapeHtmlAttr(meta.url)}">
   <meta property="og:site_name" content="${escapeHtml(meta.siteName)}">
-  ${meta.image ? `<meta property="og:image" content="${escapeHtml(meta.image)}">` : ''}
+  ${meta.image ? `<meta property="og:image" content="${escapeHtmlAttr(meta.image)}">` : ''}
   <meta name="twitter:card" content="summary_large_image">
   <meta name="twitter:title" content="${escapeHtml(meta.title)}">
   <meta name="twitter:description" content="${escapeHtml(meta.description)}">
-  ${meta.image ? `<meta name="twitter:image" content="${escapeHtml(meta.image)}">` : ''}
+  ${meta.image ? `<meta name="twitter:image" content="${escapeHtmlAttr(meta.image)}">` : ''}
   <script>
     window.location.replace('/recipes/' + ${JSON.stringify(meta.url.split('/').pop())});
   </script>
 </head>
 <body>
-  <p>Redirecting to <a href="${escapeHtml(meta.url)}">${escapeHtml(meta.title)}</a>...</p>
+  <p>Redirecting to <a href="${escapeHtmlAttr(meta.url)}">${escapeHtml(meta.title)}</a>...</p>
 </body>
 </html>`;
-
-export function escapeHtml(text: string): string {
-  return text
-    .replace(/&/g, '&amp;')
-    .replace(/</g, '&lt;')
-    .replace(/>/g, '&gt;')
-    .replace(/"/g, '&quot;')
-    .replace(/'/g, '&#39;')
-    .replace(/\//g, '&#x2F;');
-}
 
 share.get('/:slug', async (c) => {
   const slug = c.req.param('slug')!;
   try {
     const meta = await getRecipeMeta(slug);
     if (meta.visibility !== 'public') {
-      return c.html(
-        '<!DOCTYPE html><html><head><title>Not Found</title></head><body><h1>404 — Recipe not found</h1></body></html>',
-        404,
-      );
+      return c.html(RECIPE_NOT_FOUND_HTML, 404);
     }
 
     const baseUrl = Deno.env.get('APP_URL') ||
@@ -77,10 +68,7 @@ share.get('/:slug', async (c) => {
   } catch (err: unknown) {
     const message = err instanceof Error ? err.message : String(err);
     if (message === 'RECIPE_NOT_FOUND') {
-      return c.html(
-        '<!DOCTYPE html><html><head><title>Not Found</title></head><body><h1>404 — Recipe not found</h1></body></html>',
-        404,
-      );
+      return c.html(RECIPE_NOT_FOUND_HTML, 404);
     }
     throw err;
   }
