+++ compose.yml
@@ -29,6 +29,8 @@ services:
         condition: service_healthy
       mailpit:
         condition: service_started
+      garage:
+        condition: service_started
     volumes:
       - .:/app
       - ./node_modules:/app/node_modules
